projects/airflow_web_kafka_spark_postgres/src/kafka2postgres/requirements.txt
confluent-kafka==2.3.0
fastavro==1.8.4
py4j==0.10.9.7
pyspark==3.5.0
PyYAML==6.0.1
requests==2.31.0